requirements.txt
# Stock Signal Generator Dependencies
# ====================================

# Data fetching
yfinance>=0.2.40

# Data analysis
pandas>=2.0.0
numpy>=1.24.0

# Configuration
pyyaml>=6.0

# Scheduling (for automated runs)
schedule>=1.2.0
pytz>=2024.1

# Optional: Twilio for SMS (more reliable than email-to-SMS)
# twilio>=8.0.0
